+++ src/pages/index.ts
@@ -1,0 +1,3 @@
+import Todo from "./todo"
+
+export { Todo }
